requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pdf2image==1.17.0
pillow==11.2.1
PyMuPDF==1.26.0
PyPDF2==3.0.1
pytesseract==0.3.13
Werkzeug==3.1.3
